requirements.txt
numpy
PYL
pycocotools
segmentation_models_pytorch
torch
torchmetrics
albumentation
cv2